requirements.txt
beanie==1.29.0
fastapi==0.115.6
Jinja2==3.1.5
motor==3.6.1
pydantic==2.10.5
pydantic_settings==2.7.1
python-dotenv==1.0.1
uvicorn==0.34.0
